+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Features
 
 ### Enhancements
+- Use `getIsIconSideNavEnabled` API for icon side nav registration gating ([#814](https://github.com/opensearch-project/dashboards-maps/pull/814))
 
 ### Bug Fixes